requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
chromadb
PyMuPDF
pdf2image
pytesseract
Pillow
openai
google-generativeai
python-multipart
aiofiles
tqdm
duckduckgo-search
speechrecognition
pyaudio
pydub
whisper-openai
google-cloud-storage
mcp
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
